ci: adapt to new buildroot image
Similar to https://github.com/coreos/ignition/pull/1182 as a result of
https://github.com/coreos/fedora-coreos-config/pull/740.

As a bonus, we now get a clear separate stage for building. Add a stage
for running unit tests, but keep it commented out for now and we'll
circle back to it when it's cleaner to do once we have:

https://github.com/coreos/fedora-coreos-config/pull/917

GitHub Actions covers unit tests anyway.